Add Base component for Switch Responsive Design Your Shirt Page
