+++ pages/designer.vue
@@ -1,6 +1,6 @@
 <template>
     <div class="container">
-        <div class="controls-container-left">
+        <div class="controls-container-left flex-item">
             <div>
                 <ColorSwitcher></ColorSwitcher>
             </div>
@@ -8,16 +8,17 @@
         <div class="shirt-container">
             <TShirtDesigner></TShirtDesigner>
         </div>
-        <div class="controls-container-right">
-            <MotiveSwitcher></MotiveSwitcher>
+        <div class="controls-container-right flex-item">
+            <MotiveSwitchers></MotiveSwitchers>
         </div>
     </div>
 </template>
 
 <script setup lang="ts">
-import TShirtDesigner from '@/src/components/TShirtDesigner.vue'
-import ColorSwitcher from '@/src/components/ColorSwitcher.vue'
-import MotiveSwitcher from '@/src/components/MotiveSwitcher.vue'
+import ColorSwitcher from '@/src/components/switcher/ColorSwitcher.vue';
+import TShirtDesigner from '@/src/components/TShirtDesigner.vue';
+
+import MotiveSwitchers from '@/src/components/switcher/MotiveSwitcher.vue';
 </script>
 
 <style scoped>
@@ -49,4 +50,50 @@ import MotiveSwitcher from '@/src/components/MotiveSwitcher.vue'
     justify-content: center;
     align-items: center;
 }
+
+@media (max-width: 799px) {
+    .shirt-container {
+            height: 50vh;
+            flex-grow: 1;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+    }
+
+
+    .flex-item {
+        display: flex;
+        width: 100%;
+    }
+    .controls-container-right {
+        top: 70%;
+        transform: rotate(90deg);
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        height: 100px;
+    }
+
+    .controls-container-left {
+        top: 80%;
+        transform: rotate(90deg);
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        height: 100px;
+    }
+}
+
+@media (max-width: 450px) {
+    .controls-container-right,
+    .controls-container-left {
+        scale: 0.6
+    }
+
+
+}
+
+
 </style>
