+++ tests/test.spec.ts
@@ -2,5 +2,5 @@
 import { test } from 'vitest'
 
 test('my test', () => {
-  // ... test with Nuxt environment!
+    // ... test with Nuxt environment!
 })
