+++ vitest.config.ts
@@ -1,5 +1,5 @@
 import { defineVitestConfig } from '@nuxt/test-utils/config'
 
 export default defineVitestConfig({
-  // any custom Vitest config you require
+    // any custom Vitest config you require
 })
